+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Added a Changelog, that retroactively covers all versions.
+- Added an option (-v, --verbose) to enable/disable verbose output (Issue #5)
 
 ### Fixed